+++ src/main/java/com/raoulvdberge/refinedstorage/api/solderer/ISoldererRegistry.java
@@ -21,11 +21,11 @@ public interface ISoldererRegistry {
     /**
      * Returns a solderer recipe from the rows.
      *
-     * @param rows an item handler, where slots 0 - 2 are the rows
+     * @param ingredients an item handler, where slots 0 - 2 are the ingredient rows
      * @return the recipe, or null if no recipe was found
      */
     @Nullable
-    ISoldererRecipe getRecipe(@Nonnull IItemHandler rows);
+    ISoldererRecipe getRecipe(@Nonnull IItemHandler ingredients);
 
     /**
      * @return a list with all the solderer recipes